impose more consistent names to tw handlers and types, including for the fake apps used in tests
